requirements.txt
Django==4.1.3
psycopg2==2.9.5
djangorestframework==3.14.0
django-filter==22.1
dj-rest-auth==2.2.5
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2